+++ stephandlers/employeeStepTwo.ts
@@ -1,7 +1,9 @@
-import {Composer, Markup} from "telegraf";
-import {MyContext} from "../types";
-import { appendTask, updateShift, updateTask} from "../google/spreadsheet";
-import {dateString, makeButtons, trigger} from "../functions";
+import {Composer, Markup} from 'telegraf'
+import {MyContext} from '../types'
+import { appendTask, updateShift, updateTask} from '../google/spreadsheet'
+import {dateString, makeButtons, trigger} from '../functions'
+import {getDayRef, getSessionRef} from '../google/firestore'
+import {FieldValue} from 'firebase-admin/firestore'
 
 export const employeeStepTwo = new Composer<MyContext>()
 
@@ -9,20 +11,36 @@ export const employeeStepTwo = new Composer<MyContext>()
 employeeStepTwo.action(trigger, async (ctx) => {
     try {
         await ctx.answerCbQuery()
-        await ctx.deleteMessage()
         if(ctx.scene.session.taskMessage){
             await ctx.deleteMessage(ctx.scene.session.taskMessage)
+            ctx.scene.session.taskMessage = 0
         }
+        if(ctx.scene.session.lastMessage){
+            await ctx.deleteMessage(ctx.scene.session.lastMessage)
+            ctx.scene.session.lastMessage = 0
+        }
+
 
         const message = await ctx.reply('Выберите задачу из списка или просто напишите её сообщением', Markup.inlineKeyboard(makeButtons(ctx.scene.session.tasks)))
 
-
         const id = message.chat.id
+
+        const dayRef = getDayRef(id, ctx.scene.session.day)
+        const sessionRef = getSessionRef(id)
+
         if (ctx.scene.session.currentTask) {
 
             const task = ctx.scene.session.tasks[+ctx.match[0]]
 
             await updateTask(id, ctx.scene.session.currentTask?.column, ctx.scene.session.table,  ctx.scene.session.taskStart)
+            const endTime = new Date()
+            await dayRef.update('tasks', FieldValue.arrayUnion({
+                end: endTime.toLocaleString().replace(/,/, ''),
+                name: ctx.scene.session.currentTask.name,
+                start: ctx.scene.session.taskStart.toLocaleString().replace(/,/, ''),
+                time: dateString(+endTime - +ctx.scene.session.taskStart)
+            }))
+
             ctx.scene.session.taskStart = new Date()
             const column = await appendTask(id, task, ctx.scene.session.name,ctx.scene.session.table,  ctx.scene.session.taskStart)
             if (column) {
@@ -34,7 +52,10 @@ employeeStepTwo.action(trigger, async (ctx) => {
                 ctx.scene.session.taskMessage = taskMessage.message_id
             }
 
+
+
         }
+
         if (!ctx.scene.session.currentTask) {
             const task = ctx.scene.session.tasks[+ctx.match[0]]
             ctx.scene.session.taskStart = new Date()
@@ -48,9 +69,23 @@ employeeStepTwo.action(trigger, async (ctx) => {
                 ctx.scene.session.taskMessage = taskMessage.message_id
             }
 
+
         }
 
         ctx.scene.session.lastMessage = message.message_id
+        await sessionRef.update({
+            endMessage: ctx.scene.session.endMessage,
+            lastMessage: ctx.scene.session.lastMessage,
+            pauseTime: ctx.scene.session.pauseTime,
+            column: ctx.scene.session.column,
+            currentTask: ctx.scene.session.currentTask,
+            pauseStart: ctx.scene.session.pauseStart,
+            pauseMsg: ctx.scene.session.pauseMsg,
+            pauseToString: ctx.scene.session.pauseToString,
+            taskMessage: ctx.scene.session.taskMessage,
+            taskStart: ctx.scene.session.taskStart,
+        })
+
         return ctx.wizard.selectStep(1)
     }
     catch (err) {
@@ -60,17 +95,61 @@ employeeStepTwo.action(trigger, async (ctx) => {
 
 employeeStepTwo.hears(['Закончить смену'], async (ctx) => {
     try {
-        await ctx.deleteMessage(ctx.scene.session.lastMessage)
-        await ctx.deleteMessage(ctx.scene.session.endMessage)
+        if(ctx.scene.session.taskMessage){
+            await ctx.deleteMessage(ctx.scene.session.taskMessage)
+            ctx.scene.session.taskMessage = 0
+        }
+        if(ctx.scene.session.lastMessage){
+            await ctx.deleteMessage(ctx.scene.session.lastMessage)
+            ctx.scene.session.lastMessage = 0
+        }
+        if(ctx.scene.session.endMessage){
+            await ctx.deleteMessage(ctx.scene.session.endMessage)
+            ctx.scene.session.endMessage = 0
+        }
 
+
+        const pause = dateString(ctx.scene.session.pauseTime)
+        const id = ctx.message.from.id
+        const dayRef = getDayRef(id, ctx.scene.session.day)
+        const sessionRef = getSessionRef(id)
 
         await ctx.reply('До завтра', Markup.keyboard([['Начать смену']]).resize(true))
         if (ctx.scene.session.currentTask) {
             await updateTask(ctx.message.from.id, ctx.scene.session.currentTask?.column, ctx.scene.session.table,  ctx.scene.session.taskStart)
+
+            const endTime = new Date()
+            await dayRef.update('tasks', FieldValue.arrayUnion({
+                end: endTime.toLocaleString().replace(/,/, ''),
+                name: ctx.scene.session.currentTask.name,
+                start: ctx.scene.session.taskStart.toLocaleString().replace(/,/, ''),
+                time: dateString(+endTime - +ctx.scene.session.taskStart),
+                pause,
+            }))
             ctx.scene.session.currentTask = null
         }
-        const pause = dateString(ctx.scene.session.pauseTime)
+
+        if (ctx.scene.session.pauseToString !== '') {
+            await updateTask(id, ctx.scene.session.pauseToString, ctx.scene.session.table,  new Date())
+        }
+
         await updateShift(ctx.message.from.id, ctx.scene.session.column, ctx.scene.session.table, pause)
+
+
+        await dayRef.update({end: new Date().toLocaleTimeString()})
+        await sessionRef.update({
+            lastMessage: 0,
+            endMessage: 0,
+            column: '',
+            currentTask: null,
+            pauseStart: 0,
+            pauseToString: '',
+            pauseTime: 0,
+            pauseMsg: 0,
+            taskMessage: 0,
+            taskStart: 0,
+        })
+
         return ctx.scene.leave()
     }
     catch (err) {
@@ -80,10 +159,45 @@ employeeStepTwo.hears(['Закончить смену'], async (ctx) => {
 
 employeeStepTwo.hears(['Пауза'], async (ctx) => {
     try {
+        if(ctx.scene.session.taskMessage){
+            await ctx.deleteMessage(ctx.scene.session.taskMessage)
+            ctx.scene.session.taskMessage = 0
+        }
+        if(ctx.scene.session.lastMessage){
+            await ctx.deleteMessage(ctx.scene.session.lastMessage)
+            ctx.scene.session.lastMessage = 0
+        }
+        if(ctx.scene.session.endMessage){
+            await ctx.deleteMessage(ctx.scene.session.endMessage)
+            ctx.scene.session.endMessage = 0
+        }
+
+
         ctx.scene.session.pauseStart = Date.now()
-        await ctx.deleteMessage(ctx.scene.session.lastMessage)
+        const id = ctx.message.from.id
+        const sessionRef = getSessionRef(id)
+
+        const column = await appendTask(id, 'Пауза', ctx.scene.session.name, ctx.scene.session.table,  new Date())
+        if (column) {
+            ctx.scene.session.pauseToString = column
+        }
+
+
         const message = await ctx.reply('Смена на паузе', Markup.keyboard([['Закончить смену', 'Возобновить']]).resize(true))
         ctx.scene.session.lastMessage = message.message_id
+
+        await sessionRef.update({
+            endMessage: ctx.scene.session.endMessage,
+            lastMessage: ctx.scene.session.lastMessage,
+            pauseTime: ctx.scene.session.pauseTime,
+            column: ctx.scene.session.column,
+            currentTask: ctx.scene.session.currentTask,
+            pauseStart: ctx.scene.session.pauseStart,
+            pauseMsg: ctx.scene.session.pauseMsg,
+            pauseToString: ctx.scene.session.pauseToString,
+            taskMessage: ctx.scene.session.taskMessage,
+            taskStart: ctx.scene.session.taskStart ? ctx.scene.session.taskStart : new Date(),
+        })
         return ctx.wizard.selectStep(2)
     }
     catch (err) {
@@ -95,14 +209,30 @@ employeeStepTwo.on('text', async (ctx) => {
     try {
         if(ctx.scene.session.taskMessage){
             await ctx.deleteMessage(ctx.scene.session.taskMessage)
+            ctx.scene.session.taskMessage = 0
         }
-        await ctx.deleteMessage(ctx.scene.session.lastMessage)
+        if(ctx.scene.session.lastMessage){
+            await ctx.deleteMessage(ctx.scene.session.lastMessage)
+            ctx.scene.session.lastMessage = 0
+        }
+
         const message = await ctx.reply('Выберите задачу из списка или просто напишите её сообщением', Markup.inlineKeyboard(makeButtons(ctx.scene.session.tasks)))
 
         const id = message.chat.id
+        const dayRef = getDayRef(id, ctx.scene.session.day)
+        const sessionRef = getSessionRef(id)
         if (ctx.scene.session.currentTask) {
             const task = ctx.message.text
             await updateTask(id, ctx.scene.session.currentTask?.column, ctx.scene.session.table, ctx.scene.session.taskStart)
+
+            const endTime = new Date()
+            await dayRef.update('tasks', FieldValue.arrayUnion({
+                end: endTime.toLocaleString().replace(/,/, ''),
+                name: ctx.scene.session.currentTask.name,
+                start: ctx.scene.session.taskStart.toLocaleString().replace(/,/, ''),
+                time: dateString(+endTime - +ctx.scene.session.taskStart)
+            }))
+
             ctx.scene.session.taskStart = new Date()
             const column = await appendTask(id, task, ctx.scene.session.name, ctx.scene.session.table,  ctx.scene.session.taskStart)
             if (column) {
@@ -113,6 +243,7 @@ employeeStepTwo.on('text', async (ctx) => {
                 const taskMessage = await ctx.reply(`Сейчас выполняется задача: ${ctx.scene.session.currentTask.name}`)
                 ctx.scene.session.taskMessage = taskMessage.message_id
             }
+
 
         }
         if (!ctx.scene.session.currentTask) {
@@ -131,6 +262,19 @@ employeeStepTwo.on('text', async (ctx) => {
         }
 
         ctx.scene.session.lastMessage = message.message_id
+        await sessionRef.update({
+            endMessage: ctx.scene.session.endMessage,
+            lastMessage: ctx.scene.session.lastMessage,
+            pauseTime: ctx.scene.session.pauseTime,
+            column: ctx.scene.session.column,
+            currentTask: ctx.scene.session.currentTask,
+            pauseStart: ctx.scene.session.pauseStart,
+            pauseMsg: ctx.scene.session.pauseMsg,
+            pauseToString: ctx.scene.session.pauseToString,
+            taskMessage: ctx.scene.session.taskMessage,
+            taskStart: ctx.scene.session.taskStart,
+        })
+
         return ctx.wizard.selectStep(1)
     }
     catch (err) {
